Flask_JWT/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
psycopg2==2.9.3
pycryptodome==3.14.0
PyJWT==2.3.0
Werkzeug==2.0.2
